+++ NEWS.md
@@ -1,3 +1,8 @@
+# cvap 0.1.0
+* Adds support for 2020 CVAP estimates
+* Improves `cvap_distribute()` to avoid population loss, following the `geomander::estimate_down()` approach.
+* Adds `clean` argument to `cvap_distribute_censable()`
+
 # cvap 0.0.3
 
 * Added a `NEWS.md` file to track changes to the package.